refact: move code out of core pkgs
