+++ pkg/cgtproxy/error.go
@@ -1,4 +1,4 @@
-package core
+package cgtproxy
 
 import (
 	"errors"
