+++ pkg/cgtproxy/run.go
@@ -1,4 +1,4 @@
-package core
+package cgtproxy
 
 import (
 	"context"
