+++ pkg/gtproxier/error.go
@@ -1,4 +1,4 @@
-package core
+package gtproxier
 
 import (
 	"errors"
