+++ pkg/gtproxier/new.go
@@ -1,37 +1,21 @@
-package core
+package gtproxier
 
 import (
-	"context"
-
-	"github.com/black-desk/cgtproxy/internal/cgmon"
-	"github.com/black-desk/cgtproxy/internal/fswatcher"
-	"github.com/black-desk/cgtproxy/internal/routeman"
-	"github.com/black-desk/cgtproxy/pkg/cgtproxy/config"
+	"github.com/black-desk/cgtproxy/pkg/gtproxier/config"
 	. "github.com/black-desk/lib/go/errwrap"
-	"github.com/sourcegraph/conc/pool"
 	"go.uber.org/zap"
 )
 
 type Core struct {
 	cfg *config.Config
 
-	pool   *pool.ContextPool
-	log    *zap.SugaredLogger
-	stopCh chan error
-
-	components *components
-}
-
-type components struct {
-	w *fswatcher.Watcher
-	m *cgmon.Monitor
-	r *routeman.RouteManager
+	log *zap.SugaredLogger
 }
 
 type Opt = (func(*Core) (*Core, error))
 
 func New(opts ...Opt) (ret *Core, err error) {
-	defer Wrap(&err, "create new cgtproxy core")
+	defer Wrap(&err, "create new gtproxier core")
 
 	c := &Core{}
 	for i := range opts {
@@ -51,12 +35,6 @@ func New(opts ...Opt) (ret *Core, err error) {
 		return
 	}
 
-	c.pool = pool.New().
-		WithContext(context.Background()).
-		WithCancelOnError()
-
-	c.stopCh = make(chan error, 1)
-
 	ret = c
 
 	c.log.Debugw("Create a new core.",
@@ -70,7 +48,6 @@ func WithConfig(cfg *config.Config) Opt {
 	return func(core *Core) (ret *Core, err error) {
 		core.cfg = cfg
 		ret = core
-
 		return
 	}
 }
